% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayCompliance.R
\name{DayCompliance}
\alias{DayCompliance}
\title{Calculate Within Day Compliance}
\usage{
DayCompliance(json, id = NULL, coor = NULL, time = NULL,
  time.format = NULL, timebin = NULL, hour.require = NULL,
  output = c("json", "dataframe"))
}
\arguments{
\item{json}{a json object.}

\item{id}{the name of personal identifier in json.}

\item{coor}{longitude and latitude of tracking points in the format of c("lon","lat").}

\item{time}{local time variable}

\item{time.format}{format of \code{time} to pass to \code{as.POSIXct}}

\item{timebin}{time bin to test whether at least one point is recorded}

\item{hour.require}{the number of hours required for a day to be successful}

\item{output}{a string object indicating what format the output should be. Must be
either "json" or "dataframe".}
}
\value{
a data frame or a json
}
\description{
This function calculates within day compliance and returns hours completed within
a day and a completion value of 1 or 0 to indicate whether
the tracking data meets the minimum tracking requirements. Requirements are passed
in the parameters. For a day to be successful, hence \code{completion = 1}, there
must be at least one point in each time bin (i.e. 15 minutes).
}
\examples{

}
