% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafeIfelse.R
\name{SafeIfelse}
\alias{SafeIfelse}
\title{Keep Date Format in ifelse}
\usage{
SafeIfelse(cond, yes, no)
}
\arguments{
\item{cond}{a true or false condition statement which can be coerced to logical mode.}

\item{yes}{return values for true elements of \code{cond}.}

\item{no}{return values for false elements of \code{cond}.}
}
\value{
A vector of the same length and attributes (including dimensions and "class") as cond
and data values from the values of \code{yes} or \code{no}.
}
\description{
This function keeps the correct date format in a ifelse condition function
}
\examples{

}

