% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.track.R
\docType{data}
\name{sample.track}
\alias{sample.track}
\title{Sample tracks with 1000 observations}
\format{A data frame with 1000 rows and 4 variables:
\describe{
  \item{id}{unique identifier for individual}
  \item{latitude}{latitude for a geolocation}
  \item{longitude}{longitude for a geolocation}
  \item{deviceLocaltime}{local time when a geolocation is recorded, format is "%Y-%m-%dT%H:%M:%S+00:00"
  }
}}
\usage{
sample.track
}
\description{
A dataset containing 1000 sample tracks.
}
\keyword{datasets}
