% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeBin.R
\name{TimeBin}
\alias{TimeBin}
\title{Calculate Compliance Status Given a Time Bin}
\usage{
TimeBin(df, time = NULL, timebin = NULL, groupvar = NULL)
}
\arguments{
\item{df}{a data frame object.}

\item{groupvar}{grouping object to stratify time objects. Recommend to be ID for each individual.
If \code{groupvar} is not specified, the entire data frame will be considered as one group to
calculate displacement.}

\item{coor}{longitude and latitude of the spatial points in the format of c("lon","lat").}
}
\value{
a list of distance values. Unit is meters.
}
\description{
This function calculates the distance between two consecutive stay events using their geolocations.
}
\examples{
data(mobility)
mobility_stay<- stayevent(mobility, coor = c("lon","lat"), time = "datetime", dist.threshold = 100,
time.threshold = 30, time.units = "mins", groupvar = "id")
mobility_stay<- aggregate(cbind(id, stayeventlon, stayeventlat)~stayeventgroup, mobility_stay, mean)
mobility_stay$dispm<- displacement(mobility_stay, coor = c("stayeventlon","stayeventlat"), groupvar = "id")

}
