% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackJsonToDF.R
\name{TrackJsonToDF}
\alias{TrackJsonToDF}
\title{Transform Tracking Data from json to Data Frame}
\usage{
TrackJsonToDF(json, id = NULL, coor = NULL, time = NULL,
  time.format = NULL)
}
\arguments{
\item{json}{a json object.}

\item{id}{the name of personal identifier in json.}

\item{coor}{longitude and latitude of tracking points in the format of c("lon","lat").}

\item{time}{local time variable}

\item{time.format}{format of \code{time} to pass to \code{as.POSIXct}}
}
\value{
a data frame.
}
\description{
This function transforms tracking data from json to data frame and rename variables to
desired names.
}
\examples{

}

